continued updates to structure
