+++ src/plugins/plugins.go
@@ -15,7 +15,7 @@
    You should have received a copy of the GNU Affero General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-package agent
+package plugins
 
 import (
 	log "github.com/Sirupsen/logrus"
