+++ src/utils/utils.go
@@ -1,6 +1,6 @@
 /*
    PulseHA - HA Cluster Daemon
-   Copyright (C) 2017  Andrew Zak <andrew@pulseha.com>
+   Copyright (C) 2017-2018  Andrew Zak <andrew@pulseha.com>
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License as published
@@ -15,20 +15,18 @@
    You should have received a copy of the GNU Affero General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-package agent
+package utils
 
 import (
-	"errors"
-	log "github.com/Sirupsen/logrus"
-	"io/ioutil"
-	"net"
-	"os"
-	"os/exec"
-	"strings"
-	"time"
-	"strconv"
-	"runtime"
-	"github.com/Syleron/PulseHA/proto"
+"errors"
+log "github.com/Sirupsen/logrus"
+"io/ioutil"
+"net"
+"os"
+"os/exec"
+"strings"
+"time"
+"strconv"
 )
 
 /**
@@ -247,112 +245,4 @@ func WriteTextFile(contents string, file string) error {
 		return err
 	}
 	return nil
-}
-
-/**
-Networking - Bring up the groups on the current node
-*/
-func makeMemberActive() error {
-	log.Debug("Utils:MakeMemberActive() Local node now active")
-	configCopy := gconf.GetConfig()
-	for name, node := range configCopy.Nodes {
-		if name == gconf.getLocalNode() {
-			for iface, groups := range node.IPGroups {
-				for _, groupName := range groups {
-					makeGroupActive(iface, groupName)
-				}
-			}
-		}
-	}
-	return nil
-}
-
-/**
-Networking - Bring down the ip groups on the current node
- */
-func makeMemberPassive() error {
-	log.Debug("Utils:MakeMemberPassive() Local node now passive")
-	configCopy := gconf.GetConfig()
-	for name, node := range configCopy.Nodes {
-		if name == gconf.getLocalNode() {
-			for iface, groups := range node.IPGroups {
-				for _, groupName := range groups {
-					makeGroupPassive(iface, groupName)
-				}
-			}
-		}
-	}
-	return nil
-}
-
-/**
-Bring up an []ips for a specific interface
- */
-func bringUpIPs(iface string, ips []string) error {
-	plugin := pulse.Plugins.getNetworkingPlugin()
-	if plugin == nil {
-		log.Debug("No networking plugin.. skipping network action")
-		return nil
-	}
-	err := plugin.Plugin.(PluginNet).BringUpIPs(iface, ips)
-	return err
-}
-
-/**
-Bring down an []ips for a specific interface
- */
-func bringDownIPs(iface string, ips []string) error {
-	plugin := pulse.Plugins.getNetworkingPlugin()
-	if plugin == nil {
-		log.Debug("No networking plugin.. skipping network action")
-		return nil
-	}
-	err := plugin.Plugin.(PluginNet).BringDownIPs(iface, ips)
-	return err
-}
-
-/**
-
- */
-func MyCaller() string {
-	// we get the callers as uintptrs - but we just need 1
-	fpcs := make([]uintptr, 1)
-	// skip 3 levels to get to the caller of whoever called Caller()
-	n := runtime.Callers(3, fpcs)
-	if n == 0 {
-		return "n/a" // proper error her would be better
-	}
-	// get the info of the actual function that's in the pointer
-	fun := runtime.FuncForPC(fpcs[0] - 1)
-	if fun == nil {
-		return "n/a"
-	}
-	// return its name
-	return fun.Name()
-}
-
-/**
-
- */
-func setLogLevel(level string) {
-	logLevel, err := log.ParseLevel(level)
-	if err != nil {
-		panic(err.Error())
-	}
-	log.SetLevel(logLevel)
-}
-
-/**
-Determine who is the correct active node if more than one active is brought online
- */
-func getFailOverCountWinner(members []*proto.MemberlistMember) string {
-	for _, member := range members {
-		if member.Status != proto.MemberStatus_UNAVAILABLE {
-			tym, _ := time.Parse(time.RFC1123, member.LastReceived)
-			if tym == (time.Time{}) {
-				return member.Hostname
-			}
-		}
-	}
-	return ""
 }
